+++ test/libvirt/collection/network_collection_test.rb
@@ -23,6 +23,55 @@ Protest.describe("network collection") do
     end
   end
 
+  context "a new network" do
+    setup do
+      @spec = Libvirt::Spec::Network.new
+      @spec.name = "My New Network"
+    end
+
+    context "defining a new network" do
+      should "define the new network when the spec is valid" do
+        result = nil
+        assert_nothing_raised { result = @instance.define(@spec) }
+        assert result.is_a?(Libvirt::Network)
+        assert !result.active?
+        assert_equal @spec.name, result.name
+      end
+
+      should "define the new network when the spec is a string" do
+        result = nil
+        assert_nothing_raised { result = @instance.define(@spec.to_xml) }
+        assert result.is_a?(Libvirt::Network)
+        assert !result.active?
+        assert_equal @spec.name, result.name
+      end
+
+      should "raise an error when the specification is not valid" do
+        assert_raise(Libvirt::Exception::LibvirtError) {
+          @instance.define("")
+        }
+      end
+    end
+
+    context "creating a new network" do
+      should "create the new network with the specification" do
+        result = nil
+        assert_nothing_raised { result = @instance.create(@spec) }
+        assert result.is_a?(Libvirt::Network)
+        assert result.active?
+        assert_equal @spec.name, result.name
+      end
+
+      should "create the new network with the string spec" do
+        result = nil
+        assert_nothing_raised { result = @instance.create(@spec.to_xml) }
+        assert result.is_a?(Libvirt::Network)
+        assert result.active?
+        assert_equal @spec.name, result.name
+      end
+    end
+  end
+
   should "provide a list of active networks" do
     active = @instance.active
     assert active.is_a?(Array)